Add --ignore=conflictingarchall
This is useful if autobuilders for more than one architecture will build
Architecture: all packages of the same version. For instance, if amd64
and i386 builders both try to build dbus, and the amd64 builder finishes
first, the amd64-built version of dbus-1-doc will be included in the
archive and the i386-built version will be discarded.

Based on a patch by Sjoerd Simons.

[Added commit message, fixed a warning message -Simon McVittie]
Link: http://bugs.debian.org/697630