Add sha512 implementation
This public domain implementation stems from
https://akkadia.org/drepper/SHA-crypt.txt.

The applicable changes that Bernhard made for the SHA256 implementation
are copied to this version so that the code is consistent.

The changes in f5228d7718 ("fix bug in sha256 calculation of very large files")
and 2688a8902 ("fix undefined memory access in sha256") are not copied because
they do not seem to be applicable. Also, the merge of the split "total" counter
is not copied.